% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wd.R
\name{dist.cohen.d}
\alias{dist.cohen.d}
\title{dist.cohen.d - cohen.d statistic}
\usage{
dist.cohen.d(dm, f)
}
\arguments{
\item{dm}{A distance matrix}

\item{f}{factor}
}
\description{
The dist.cohen.d statistic is based on a matrix and a factor
}
