% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wd.R
\name{dist.group.sigma2}
\alias{dist.group.sigma2}
\title{dist.group.sigma2 - group sigma2 distance measure}
\usage{
dist.group.sigma2(dm, f)
}
\arguments{
\item{dm}{A distance matrix}

\item{f}{factor}
}
\description{
The dist.group.sigma2 statistic is based on a matrix and a factor
}
