% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wd.R
\name{dist.sigma2}
\alias{dist.sigma2}
\title{dist.sigma2 - sigma2 distance measure}
\usage{
dist.sigma2(dm)
}
\arguments{
\item{dm}{A distance matrix}
}
\description{
The dist.sigma2 statistic is based on a matrix and a factor
}
