% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wd.R
\name{dist.ss2}
\alias{dist.ss2}
\title{dist.ss2 - ss2 distance measure}
\usage{
dist.ss2(dm2, f)
}
\arguments{
\item{dm2}{A distance matrix of square distances}

\item{f}{factor}
}
\description{
The dist.ss2 statistic is based on a matrix and a factor
}
