% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wd.R
\name{Tw2.posthoc.1vsAll.tests}
\alias{Tw2.posthoc.1vsAll.tests}
\title{Tw2.posthoc.1vsAll.tests}
\usage{
Tw2.posthoc.1vsAll.tests(dm, f, nrep = 999, strata = NULL)
}
\arguments{
\item{dm}{A distance matrix}

\item{f}{factor}

\item{nrep}{factor}

\item{strata}{factor}
}
\description{
Run Tw2.test one or more times, with an optional stratification factor
}
