% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wd.R
\name{Tw2.test}
\alias{Tw2.test}
\title{Tw2.test - Tw2.test - run permuations of Tw2 statistic}
\usage{
Tw2.test(dm, f, nrep = 999, strata = NULL)
}
\arguments{
\item{dm}{A distance matrix description of the parameter 'x'. The
description can span multiple lines.}

\item{f}{factor.}

\item{nrep}{number of times to run the test.}

\item{strata}{the variable to use to stratify}
}
\value{
returns a list with class \strong{htest} containing

method	- description of test

statistic - observed value of the test statistic

p.value - approximate p-value of the test

nrep - number of repetitions
}
\description{
The Tw2.test calculates the statistic for a given number of reps
is based on a matrix and a factor
}
\details{
An optional strata can be specified.
}
