% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wd.R
\name{WdS}
\alias{WdS}
\title{This is the title.}
\usage{
WdS(dm, f)
}
\arguments{
\item{dm}{A description of the parameter 'x'. The
description can span multiple lines.}

\item{f}{A description of the parameter 'y'.}
}
\description{
This is the description.
}
\details{
These are further details.
}
\section{A Custom Section}{


Text accompanying the custom section.
}

\examples{
WdS(1, 2) ## returns 3

## don't run this in calls to 'example(add_numbers)'
\dontrun{
   WdS(2, 3)
}

## don't test this during 'R CMD check'
\donttest{
   WdS(4, 5)
}
}
