% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wd.R
\name{WdS.test}
\alias{WdS.test}
\title{WdS.test - run WdS one or more times}
\usage{
WdS.test(dm, f, nrep = 999, strata = NULL)
}
\arguments{
\item{dm}{A distance matrix}

\item{f}{factor}

\item{nrep}{factor}

\item{strata}{factor}
}
\value{
returns a list with class \strong{htest} containing

method	- description of test

statistic - observed value of the test statistic

p.value - approximate p-value of the test

nrep - number of repetitions
}
\description{
Run WdS one or more times, with an optional stratification factor
}
